+++ actix-connect/tests/test_connect.rs
@@ -2,7 +2,7 @@ use std::io;
 
 use actix_codec::{BytesCodec, Framed};
 use actix_rt::net::TcpStream;
-use actix_service::{service_fn, Service, ServiceFactory};
+use actix_service::{fn_service, Service, ServiceFactory};
 use actix_testing::TestServer;
 use bytes::Bytes;
 use futures::SinkExt;
@@ -14,7 +14,7 @@ use actix_connect::Connect;
 #[actix_rt::test]
 async fn test_string() {
     let srv = TestServer::with(|| {
-        service_fn(|io: TcpStream| {
+        fn_service(|io: TcpStream| {
             async {
                 let mut framed = Framed::new(io, BytesCodec);
                 framed.send(Bytes::from_static(b"test")).await?;
@@ -33,7 +33,7 @@ async fn test_string() {
 #[actix_rt::test]
 async fn test_rustls_string() {
     let srv = TestServer::with(|| {
-        service_fn(|io: TcpStream| {
+        fn_service(|io: TcpStream| {
             async {
                 let mut framed = Framed::new(io, BytesCodec);
                 framed.send(Bytes::from_static(b"test")).await?;
@@ -51,7 +51,7 @@ async fn test_rustls_string() {
 #[actix_rt::test]
 async fn test_static_str() {
     let srv = TestServer::with(|| {
-        service_fn(|io: TcpStream| {
+        fn_service(|io: TcpStream| {
             async {
                 let mut framed = Framed::new(io, BytesCodec);
                 framed.send(Bytes::from_static(b"test")).await?;
@@ -75,7 +75,7 @@ async fn test_static_str() {
 #[actix_rt::test]
 async fn test_new_service() {
     let srv = TestServer::with(|| {
-        service_fn(|io: TcpStream| {
+        fn_service(|io: TcpStream| {
             async {
                 let mut framed = Framed::new(io, BytesCodec);
                 framed.send(Bytes::from_static(b"test")).await?;
@@ -100,7 +100,7 @@ async fn test_uri() {
     use std::convert::TryFrom;
 
     let srv = TestServer::with(|| {
-        service_fn(|io: TcpStream| {
+        fn_service(|io: TcpStream| {
             async {
                 let mut framed = Framed::new(io, BytesCodec);
                 framed.send(Bytes::from_static(b"test")).await?;
@@ -121,7 +121,7 @@ async fn test_rustls_uri() {
     use std::convert::TryFrom;
 
     let srv = TestServer::with(|| {
-        service_fn(|io: TcpStream| {
+        fn_service(|io: TcpStream| {
             async {
                 let mut framed = Framed::new(io, BytesCodec);
                 framed.send(Bytes::from_static(b"test")).await?;